+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 - PR #4082 Rework MultiColumns in cuDF
 - PR #4149 Use "type-serialized" for pickled types like Dask
 - PR #4171 Split java host and device vectors to make a vector truly immutable
+- PR #4167 Port `search` to libcudf++ (support multi-column searchsorted)
 - PR #4163 Assert Dask CUDA serializers have `Buffer` frames
 - PR #4165 List serializable classes once
 - PR #4168 IO readers: do not create null mask for non-nullable columns